chore: check not found upload chart tgz

+++ .github/utils/utils.sh
@@ -605,8 +605,12 @@ delete_charts_index() {
 get_incremental_chart_package() {
     for filePath in $( git diff --name-only HEAD HEAD^ ); do
         if [[ "$filePath" == "upload-charts/"*".tgz" ]]; then
-            echo "cp $filePath .cr-release-packages"
-            cp $filePath .cr-release-packages
+            if [[ -f "$filePath" ]]; then
+                echo "cp $filePath .cr-release-packages"
+                cp $filePath .cr-release-packages
+            else
+                echo "not found upload chart tgz $filePath"
+            fi
         fi
     done
 }
